requirements.txt
flask
flask-wtf
flask-limiter
flask-talisman
python-escpos
bcrypt
python-dotenv
Pillow
werkzeug
gunicorn
pymysql
cryptography
requests
